requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.33
alembic==1.13.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
jinja2==3.1.6
python-multipart==0.0.20
